requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
mysql-connector-python==8.0.33
Flask-RESTful==0.3.9
python-dotenv==1.0.0
gunicorn